requirements.txt
aiogram
python-dotenv
pydantic
requests
google-api-python-client
google-auth-httplib2
google-auth-oauthlib